Add files via upload
significant progress made on the assignment today. prior probabilities have been finished, evidential probabilities is making progress but encountered a few issues with the for loops. added another method to the learning model which sorts the data into its respective category. all in all a big step closer to the end result