Add files via upload
Today I focused on understanding how I would put the pieces together for the Bayes theorem formula. I made a large comment in the learningmodel classwhich is a step by step procedure to making an algorithm that will run the formula on the data provided. It is more pseudocode than actual code as it will act as a template for when I am making the functions for the other data points.