Add files via upload
Today I mostly finished the learning model and the algorithm now successfully trains itself dynamically on the data. I created a new dataAnalysis file where I will be testing the algorithm on the remaining data. I managed to get some of the testing done today. I also added some new functionalities to the GUI so it can run the code using the buttons.